requirements.txt
aiodns>=1.1  # aiodns speeds up DNS resolving
aiofile>=2.0.0
aiohttp==3.9.3
cchardet
cython
order_book==0.6.0
pyyaml
requests>=2.18.4
uvloop
websockets>=7.0
yapic.json>=1.6.3
redis==5.0.2
python-dotenv==1.0.1
